rename components and fix bug with weather interval
