--- web/src/views/PlayerLeaderboard.vue
@@ -1,17 +1,0 @@
-<template>
-  <div class="home">
-    <PlayerLeaderboardComponent/>
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import PlayerLeaderboardComponent from '../components/PlayerLeaderboardComponent.vue';
-
-export default {
-  name: 'leaderboard',
-  components: {
-    PlayerLeaderboardComponent,
-  },
-};
-</script>
