+++ web/src/views/Standings.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="home">
+    <StandingsComponent/>
+  </div>
+</template>
+
+<script>
+import StandingsComponent from '../components/StandingsComponent.vue';
+
+export default {
+  name: 'Standings',
+  components: {
+    StandingsComponent,
+  },
+};
+</script>
